refactored SimpleCrmMenu to contain main template creation (so adding a new main template only needs to happen in one place)
